Updates
Added a settings button, map, and the medication tab, and added more clarity to the server.R file. Also added account types to the account creation UI and the database, and moved the main UI from the renderUI function in the server file to a separate variable in the global file to reduce clutter.